requirements.txt
Flask==3.0.0
numpy==1.26.2
Pillow==10.1.0
Pillow==10.1.0
scikit_learn==1.3.2
Werkzeug==3.0.1
